Rework activity design
