+++ src/activitiesData.js
@@ -1,0 +1,28 @@
+const activities = [
+  {
+    name: "Running",
+    temp: [40, 70],
+    wind: [0, 10],
+    rain: [0, 30],
+    snow: [0, 30],
+    humidity: [0, 60],
+  },
+  {
+    name: "Skiing",
+    temp: [20, 60],
+    wind: [0, 15],
+    rain: [0, 10],
+    snow: [50, 100],
+    humidity: [0, 100],
+  },
+  {
+    name: "Fishing",
+    temp: [40, 70],
+    wind: [0, 10],
+    rain: [0, 30],
+    snow: [0, 30],
+    humidity: [0, 100],
+  },
+];
+
+export default activities;
