--- src/components/Activities/Activities.css
@@ -1,11 +1,0 @@
-.activities-parent {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  width: 90%;
-  max-width: 1200px;
-  padding: 0 1rem;
-  border-radius: 4px;
-  flex-grow: 1;
-  overflow-y: auto;
-}
