+++ src/components/Header/Header.css
@@ -17,7 +17,7 @@ header {
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 0.8rem;
+  padding: 0.8rem 1.2rem;
   width: 100%;
   background-color: #22AED1;
 }
